requirements.txt
streamlit==1.35.0
anthropic>=0.40.0
snowflake-connector-python==3.10.1
python-dotenv==1.0.1
pandas==2.2.2
sqlparse==0.5.0
pyyaml==6.0.1
